Move some things to ES6
